Refactoring codes, packages
