+++ app/src/main/java/com/example/answer/github/ui/GithubViewPagerAdapter.kt
@@ -1,8 +1,10 @@
-package com.example.answer.github.view.adapter
+package com.example.answer.github.ui
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
+import com.example.answer.github.ui.GithubListAdapter
+import com.example.answer.github.ui.PagingAdapter
 import com.example.answer.github.view.LikeFragment
 import com.example.answer.github.view.SearchFragment
 import com.example.answer.github.view.GithubActivity
