+++ app/src/main/java/com/example/answer/github/view/LikeFragment.kt
@@ -12,7 +12,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.answer.R
 import com.example.answer.databinding.FragmentLikeBinding
 import com.example.answer.github.viewmodel.GithubViewModel
-import com.example.answer.github.view.adapter.GithubListAdapter
+import com.example.answer.github.ui.GithubListAdapter
 
 class LikeFragment : Fragment() {
     private var githubViewModel: GithubViewModel? = null
